fix(Countdown): make bare-date regression test timezone-independent
Extract parseConfigDate to a colocated utils.ts (avoids the
react-refresh/only-export-components lint warning from exporting a
non-component from Widget.tsx) and add a direct unit test asserting its
local-noon anchoring via Date field getters — this doesn't depend on
process.env.TZ reassignment being honoured mid-process, unlike the
existing CountdownWidget integration test. Also switch that integration
test to vi.stubEnv/vi.unstubAllEnvs, Vitest's own env-stubbing API,
which is more reliable than direct process.env mutation.

Addresses claude[bot] review feedback on PR #2227: the original
TZ-override test could pass on unfixed code if V8/ICU ignores a
mid-process TZ reassignment.
